Changed plugin definitions.

+++ test/listedpixelselectiontest.cpp
@@ -37,7 +37,6 @@ int main()
 		Roi<int> roi("roi");
 		ListedPixelSelection<float> lps("lps");
 		std::vector<Pixel<float>*> list;
-
 		roi.left = 29;
 		roi.right = 44;
 		roi.top = 0;
@@ -46,12 +45,9 @@ int main()
 		roi.back = 1;
 		roi.before = 0;
 		roi.after = 1;
-
 		rd.filename = PENGUINFILE;
-
 		lps.range.connect(&roi.out);
 		lps.sequence.connect(&rd.out);
-
 		lps.execute();
 		list = lps.getListOfPixel();
 
@@ -89,7 +85,6 @@ int main()
 		roit.after = 1;
 		lpst.range.connect(&roit.out);
 		lpst.sequence.connect(&rdt.out);
-
 		lpst.execute();
 		std::vector<Pixel<float>*> listt = lpst.getListOfPixel();
 		Pixel<float> pix = *(listt[0]);
@@ -105,16 +100,20 @@ int main()
 				std::vector<float> intensi;
 				for (unsigned int it = 0; it < temp.getIntensity().size(); it++)
 				{
-					intensi.push_back(rdt.out()(it, temp.getX(), temp.getY(),
-							temp.getZ(), temp.getT()));
+					intensi.push_back(rdt.out()(it, (unsigned int) temp.getX(),
+							(unsigned int) temp.getY(),
+							(unsigned int) temp.getZ(),
+							(unsigned int) temp.getT()));
 				}
 				for (unsigned int iter = 0; iter < temp.getIntensity().size(); iter++)
 				{
 					if (intensi[iter] != temp.getIntensity()[iter])
 					{
 						std::cout << "original intensity: " << rdt.out()[0](
-								temp.getX(), temp.getY(), temp.getZ(),
-								temp.getT()) << std::endl;
+								(unsigned int) temp.getX(),
+								(unsigned int) temp.getY(),
+								(unsigned int) temp.getZ(),
+								(unsigned int) temp.getT()) << std::endl;
 						std::cout << "saved Pixel: "
 								<< temp.getIntensity()[iter] << std::endl;
 						std::cout << "pixel (" << temp.getX() << ", "
